Refactored platform headers
Made a centralized header, platformAPI.h to ensure cohesion in definitions for window creation and updating.

Updating main.cpp to reflect this change.

Now, the platformAPI.h header will import either windowsAPI.h or linuxAPI.h through a preprocessor macro, based on the os during compilation.

The platformAPI.h file is imported by main, as is, so on compile, the appropriate platform api header and functions will be linked through the platformAPI header.